backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.1
loguru==0.7.2

numpy==1.26.4
Pillow==10.3.0
opencv-python==4.9.0.80
requests==2.32.2
aiofiles==23.2.1

torch>=2.2.0
torchvision>=0.17.0
transformers==4.41.2
sentencepiece==0.2.0
tokenizers==0.19.1
accelerate==0.30.1
safetensors==0.4.3
onnxruntime==1.18.0

manga-ocr==0.1.12
easyocr==1.7.1
pytesseract==0.3.10

diffusers==0.28.2
controlnet-aux==0.0.9

# NOTE: keep lama-cleaner out of the base lock because available wheels pin
# diffusers==0.16.1 and conflict with the SDXL path (diffusers==0.28.2).
# Install it in an isolated env only when its Flask service is explicitly needed.
basicsr==1.4.2
realesrgan==0.3.0

sqlalchemy==2.0.30
scikit-learn==1.5.0
faiss-cpu==1.8.0
diskcache==5.6.3

huggingface-hub==0.23.2

pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
